chore: move rules and variables to themes/teams
