+++ src/components/Button/Button.tsx
@@ -2,8 +2,8 @@ import * as PropTypes from 'prop-types'
 import * as React from 'react'
 
 import { UIComponent, childrenExist, customPropTypes } from '../../lib'
-import buttonRules from './buttonRules'
-import buttonVariables from './buttonVariables'
+import buttonRules from '../../themes/teams/components/Button/buttonRules'
+import buttonVariables from '../../themes/teams/components/Button/buttonVariables'
 import Icon from '../Icon'
 import Text from '../Text'
 import { ButtonBehavior } from '../../lib/accessibility'
