+++ src/components/Icon/Icon.tsx
@@ -2,8 +2,8 @@ import * as React from 'react'
 import * as PropTypes from 'prop-types'
 import { customPropTypes, UIComponent, createShorthandFactory } from '../../lib'
 
-import iconRules from './iconRules'
-import iconVariables from './iconVariables'
+import iconRules from '../../themes/teams/components/Icon/iconRules'
+import iconVariables from '../../themes/teams/components/Icon/iconVariables'
 import svgIcons from './svgIcons'
 
 export type IconXSpacing = 'none' | 'before' | 'after' | 'both'
