+++ src/components/List/List.tsx
@@ -4,8 +4,8 @@ import * as PropTypes from 'prop-types'
 
 import { customPropTypes, UIComponent, childrenExist } from '../../lib'
 import ListItem from './ListItem'
-import listRules from './listRules'
-import listVariables from './listVariables'
+import listRules from '../../themes/teams/components/List/listRules'
+import listVariables from '../../themes/teams/components/List/listVariables'
 import { ListBehavior } from '../../lib/accessibility'
 
 class List extends UIComponent<any, any> {
