+++ src/themes/teams/components/Accordion/accordionContentRules.ts
