+++ src/themes/teams/components/Accordion/accordionContentVariables.ts
