+++ src/themes/teams/components/Accordion/accordionRules.ts
