+++ src/themes/teams/components/Accordion/accordionTitleRules.ts
