+++ src/themes/teams/components/Avatar/avatarRules.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 import { PositionProperty } from 'csstype'
 
 const getAvatarDimension = (size: number) => {
