+++ src/themes/teams/components/Avatar/avatarVariables.ts
