+++ src/themes/teams/components/Button/buttonVariables.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 
 export interface IButtonVariables {
   height: string
