+++ src/themes/teams/components/Chat/chatMessageRules.ts
