+++ src/themes/teams/components/Chat/chatMessageVariables.ts
