+++ src/themes/teams/components/Chat/chatRules.ts
