+++ src/themes/teams/components/Divider/dividerVariables.ts
