+++ src/themes/teams/components/Icon/fontAwesomeIconRules.ts
