+++ src/themes/teams/components/Image/imageVariables.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 
 export default () => ({
   width: undefined,
