+++ src/themes/teams/components/Input/inputVariables.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 
 export default (siteVars: any) => {
   const vars: any = {}
