+++ src/themes/teams/components/Label/labelVariables.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 
 export default () => {
   const vars: any = {}
