+++ src/themes/teams/components/List/listItemRules.ts
@@ -1,5 +1,5 @@
-import { pxToRem } from '../../lib'
-import { debugRoot } from '../../styles/debugRules'
+import { pxToRem } from '../../../../lib'
+import { debugRoot } from '../../../../styles/debugRules'
 
 const listItemRules = {
   root: ({ props, variables }) => {
