+++ src/themes/teams/components/List/listRules.ts
@@ -1,4 +1,4 @@
-import { debugRoot } from '../../styles/debugRules'
+import { debugRoot } from '../../../../styles/debugRules'
 
 // ----------------------------------------
 // Root
