+++ src/themes/teams/components/List/listVariables.ts
@@ -1,4 +1,4 @@
-import { pxToRem } from '../../lib'
+import { pxToRem } from '../../../../lib'
 
 export default siteVars => {
   const vars: any = {}
